pyproject.toml
[project]
name = "rag-project-kernal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.12",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-ollama>=0.3.3",
    "ollama>=0.5.1",
    "pypdf>=5.6.0",
    "sentence-transformers>=4.1.0",
]
